requirements.txt
Django
djangorestframework
numpy
requests
torch
transformers
gunicorn
asgiref
nltk
python-dotenv